The Heartbeat operation now properly releases and has sound loop logic
* The entire heartbeat loop is now even more complicate though. Thought
  it would get easier
* Still missing some cleanup logic. @hen the service fails to send to
  the claim to the client, that lock should be released.

+++ pkg/models/api/locker/v1/lock.go
@@ -21,7 +21,7 @@ type Lock struct {
 	TimeTillExipre time.Duration
 
 	// LocksHeldOrWaiting show how many clients are all trying to obtaiin the same lock
-	LocksHeldOrWaiting int
+	LocksHeldOrWaiting int64
 }
 
 //	RETURNS:
